fix: resolve compilation warnings and improve code structure
1. Fix order of member variable initialization in qw_interface
constructor
2. Move add_compile_options after ASAN flags to ensure proper priority
3. Fix member variable order in TreelandConfig to match declaration
4. Add [[maybe_unused]] attribute to unused function parameters
5. Fix surface wrapper bitfield member ordering
6. Fix cursor image index comparison type mismatch
7. Fix various minor code style and structure issues

fix: 解决编译警告并改进代码结构

1. 修复 qw_interface 构造函数中成员变量的初始化顺序
2. 将 add_compile_options 移至 ASAN 标志之后以确保优先级
3. 修复 TreelandConfig 中成员变量顺序以匹配声明
4. 为未使用的函数参数添加 [[maybe_unused]] 属性
5. 修复 surface wrapper 位域成员顺序
6. 修复光标图像索引比较的类型不匹配问题
7. 解决其他次要的代码风格和结构问题

+++ waylib/src/server/kernel/wxcursorimage.cpp
@@ -78,7 +78,7 @@ bool WXCursorImage::jumpToImage(int imageNumber)
     if (!d->cursor)
         return false;
 
-    if (imageNumber >= d->cursor->image_count)
+    if (imageNumber >= static_cast<int>(d->cursor->image_count))
         return false;
     d->currentImageNumber = imageNumber;
